requirements.txt
streamlit
pyzbar
opencv-python
Pillow
numpy
face_recognition